pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pyenvplus"
version = "0.0.1"
description = "A lightweight Python package for managing environment variables with advanced features."
readme = "README.md"
requires-python = ">=3.6"
license = {text = "MIT"}
authors = [
  {name = "Viswavr", email = "viswavr54@gmail.com"},
]
keywords = [
    "dotenv",
    "env loader",
    "type-safe environment variables",
    "configuration",
    "environment manager",
    "python env",
    "settings",
    "config",
    "environment",
    "variables",
    "12-factor",
    "development",
    "production",
    "secrets",
    "parsing",
    "casting",
    "validation",
    "hot-reload",
    "json",
    "list",
    "boolean",
    "integer",
    "float",
    "alias",
    "strict mode",
    "devops",
    "deployment",
    "flask",
    "django",
    "fastapi",
    "microservices",
    "pipeline",
    "cli",
    "automation",
    "scripting",
    "cross-platform",
    "env-var",
    "env-file",
    "env-parsing",
    "config-management",
    "settings-management",
    "application-config",
    "app-settings",
    "env-vars",
    "environment-variables",
    "python-dotenv",
    "env-plus",
    "envplus",
    "configuration-management",
    "typed-env",
    "smart-env",
    "easy-env",
    "simple-env",
    "robust-env",
    "secure-env",
    "dynamic-env",
    "reload-env",
    "auto-env",
    "env-utils",
    "env-tools",
    "env-helper",
    "env-wrapper",
    "env-handler",
    "env-processor",
    "env-reader",
    "env-writer",
    "env-access",
    "env-interface",
    "env-api",
    "env-library",
    "env-package",
    "env-module",
    "python-configuration",
    "python-settings",
    "python-environment",
    "python-variables",
    "python-config",
    "python-utils",
    "python-tools",
    "python-helpers",
    "python-library",
    "python-package",
    "python-module",
    "developer-tools",
    "dev-tools",
    "productivity",
    "utility",
    "library",
    "package",
    "module",
    "framework-agnostic",
    "lightweight",
    "fast",
    "simple",
    "easy",
    "powerful",
    "flexible",
    "extensible",
    "maintainable",
    "reliable",
    "stable",
    "secure",
    "safe",
    "typed",
    "validated",
    "modern",
    "clean",
    "python3",
    "python3.6",
    "python3.7",
    "python3.8",
    "python3.9",
    "python3.10",
    "python3.11",
    "python3.12",
]
dependencies = [
  "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["pyenvplus"]
